az-func-audio/requirements.txt
openai==1.72.0
python-dotenv==1.0.1
requests
reportlab==4.2.5
azure-functions==1.23.0
azure-cosmos==4.9.0
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
azure-storage-blob==12.25.1